\name{Prots4Prots-package}
\alias{Prots4Prots}
\alias{Prots4Prots-package}
\docType{package}
\title{
Protocols for Proteomics
}
\description{
A package comparing different statistical tools in order to find differentially expressed proteins. 
The R folder contains the toolbox, tools brought together by topic. 
The inst folder contains use cases of the toolbox.

Quantitative Proteomics

This package divides a normal analysis pipeline into multiple steps.
\tabular{ll}{
Loading data.
Treating missing values
Pre-processing
 Normalization
 Summarization
Statistics
Multiple testing correction
Differential expression determination
Inclusion of other sources of information.
}
Each step of the pipeline should be able to generate a report paragraph, itself transformed in the end into a HTML report. 
}
\details{
\tabular{ll}{
Package: \tab Prots4Prots\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-12\cr
License: \tab \cr
}
}
\author{
Celine Hernandez <Celine.Hernandez@unil.ch>
}
\references{
Bioconductor case studies
}
\keyword{ package }
